pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opinionforge"
version = "2.0.0"
description = "AI-powered editorial craft engine for generating opinion pieces with precise rhetorical control."
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.24",
    "trafilatura>=1.6",
    "anthropic>=0.30",
    "openai>=1.0",
    "tavily-python>=0.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "sse-starlette>=1.6",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19",
    "platformdirs>=4.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
desktop = [
    "pystray>=0.19",
    "Pillow>=10.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "ruff>=0.1",
    "mypy>=1.5",
    "coverage>=7.0",
]

[project.scripts]
opinionforge = "opinionforge.__main__:main"
